Relocate e2e/flowtest and e2e/lifecycle tests
