+++ tests/functional/object/mcg/lifecycle/test_mcg_namespace_lifecyle_rpc.py
