+++ tests/functional/object/mcg/lifecycle/test_mcg_namespace_s3_ops_crd.py
